+++ src/main/scala/com/atomist/source/RepoArtifactSourceLocator.scala
@@ -1,0 +1,8 @@
+package com.atomist.source
+
+trait RepoArtifactSourceLocator
+  extends ArtifactSourceLocator
+    with RepoBranchAndPath {
+
+  def repo: String
+}