+++ src/main/scala/com/atomist/source/RepoBranchAndPath.scala
@@ -1,0 +1,8 @@
+package com.atomist.source
+
+trait RepoBranchAndPath {
+
+  def path: String
+
+  def branch: String
+}