+++ src/main/scala/com/atomist/source/git/GitArtifactSourceIdentifier.scala
@@ -1,0 +1,6 @@
+package com.atomist.source.git
+
+import com.atomist.source.RepoArtifactSourceLocator
+
+trait GitArtifactSourceIdentifier extends RepoArtifactSourceLocator
+