+++ src/main/scala/com/atomist/source/git/GitArtifactSourceLocator.scala
@@ -1,0 +1,5 @@
+package com.atomist.source.git
+
+import com.atomist.source.RepoArtifactSourceLocator
+
+trait GitArtifactSourceLocator extends RepoArtifactSourceLocator