+++ CHANGELOG.md
@@ -5,8 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## [Unreleased]
+## [0.19.0]
 
+### Changed
+
+-   Added RepoArtifactSourceLocator, RepoBranchAndPath, GitArtifactSourceLocator,
+    and GitArtifactSourceIdentifierwhich were previously in the 
+    github-artifact-source repo
+    
 [Unreleased]: https://github.com/atomist/artifact-source/compare/0.14.4...HEAD
 
 ## [0.14.4]